+++ CHANGELOG.rst
@@ -3,6 +3,8 @@ Version explained:
  - minor : increase on risk of code breakage during a major release
  - bug   : increase on bug or incremental changes
 
+Version 4.2.21
+
 Version 4.2.20
  * Fix: correctly filter routes announced by the API to the right peer #1005
  * Feature: healthcheck neighbor filtering and path-information backport of #1098 and #1099